Moved over to vite for build & live reload
