+++ index.html
@@ -4,8 +4,8 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Task Manager App</title>
-    <link href="./styles.css" rel="stylesheet" />
   </head>
+
   <body class="bg-gray-100 min-h-screen flex flex-col items-center p-4">
     <main class="w-full max-w-md bg-white rounded-lg shadow p-6">
       <h1 class="text-2xl font-bold mb-4">Task Manager</h1>
@@ -27,6 +27,6 @@
       </form>
       <section id="task-list" class="mt-6 flex flex-col gap-2"></section>
     </main>
-    <script src="./app.js"></script>
+    <script type="module" src="/src/main.js"></script>
   </body>
 </html>
