+++ package.json
@@ -1,18 +1,17 @@
 {
-    "name": "task-manager-app",
-    "version": "1.0.0",
-    "description": "",
-    "main": "index.js",
-    "scripts": {
-        "build:css": "tailwindcss build src/styles.css -o public/styles.css",
-        "start": "live-server public"
-    },
-    "author": "",
-    "license": "ISC",
-    "devDependencies": {
-        "autoprefixer": "^10.4.21",
-        "live-server": "^1.2.2",
-        "postcss": "^8.5.3",
-        "tailwindcss": "^3.4.17"
-    }
+  "name": "task-manager",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
+  },
+  "devDependencies": {
+    "autoprefixer": "^10.4.21",
+    "postcss": "^8.5.3",
+    "tailwindcss": "^3.4.17",
+    "vite": "^6.3.1"
+  }
 }
