+++ tailwind.config.js
@@ -1,11 +1,11 @@
-/** @type {import('tailwindcss').Config} */
-module.exports = {
-    content: [
-      "./public/**/*.html",
-      "./src/**/*.{js,ts}", // include your JS/TS files in src/
-    ],
-    theme: {
-      extend: {},
-    },
-    plugins: [],
-  }
+// tailwind.config.js
+export default {
+  content: [
+    "./index.html",
+    "./src/**/*.{js,ts}"
+  ],
+  theme: {
+    extend: {},
+  },
+  plugins: [],
+}
