+++ vite.config.js
@@ -1,0 +1,3 @@
+import { defineConfig } from 'vite'
+
+export default defineConfig({})
